fix: Return 404 when feature or experiment not found in some routes (#107)
Update the following routes to return a 404 response with no content when the feature or experiment with the key given in the URL path doesn't exist:


GET /features/featureKey
GET /experiments/experimentKey
GET /users/userId/features/featureKey
GET /users/userId/experiments/experimentKey
POST /users/userId/features/featureKey
POST /users/userId/experiments/experimentKey

This is implemented with two new methods of CachedOptlyMiddleware in the middleware package: FeatureCtx and ExperimentCtx.

In addition to returning 404 early, when the entity is found, it's set in request context for handlers later in the chain. The final handlers for the above routes are updated to access the entity through request context, instead of directly accessing URL params.

Also changed:
Change response status from 422 to 500 in all handlers that rely on middleware to populate OptlyClient or OptlyContext, for cases when the context failed to be set properly. This is for consistency with the experiment and feature handlers updated, and the FeatureCtx and ExperimentCtx middlewares.