docs(openapi): Fix documented response code
The current API spec is out of date with the existing implementation. `/track` returns an HTTP 200 and not a 204. The change was made as part of [this](https://github.com/optimizely/agent/commit/70300ade293bfb25286fb8f619fb656274b8f4bc#diff-792ff1ebb42545c5d3e4b34233906b27L75-R70) commit, but the spec was not updated.