fix: Allowed hosts middleware does not check port (#271)
Summary:

This fixes an issue with the allowed hosts middleware that could prevent legitimate requests from being served.

The middleware extracts a host value from requests to compare agains the allow-list configured in server.allowedHosts. These request host values can take two possible forms:

1. Just a host (i.e. example.com), or
2. A host + port combination, separated by : (example.com:8080).

Previously, when the request host value did contain a port (the 2nd form), and that port differed from the server.port configuration value, the middleware would reject the request.

With this change, only the host is used for comparison against the configured allow-list. If the port is present, it is ignored.

Test Plan:

- Updated unit tests
- Manually tested